Refactor the Methods API
