+++ mollie/api/resources/__init__.py
@@ -1,7 +1,9 @@
+from .methods import Methods
 from .payments import Payments
 from .refunds import OrderRefunds, PaymentRefunds, Refunds, SettlementRefunds
 
 __all__ = [
+    Methods,
     OrderRefunds,
     Payments,
     PaymentRefunds,
